[StatsPlugin] Plumb channel args through `StatsPluginChannelScope`
This allows CallTracers to be created with parameters dictated by
channel args.

For the moment, I've used the EventEngine `EndpointConfig` API to expose
the channel args here, so as to avoid directly exposing
`grpc_core::ChannelArgs`.  We should determine a better API here before
we de-experimentalize the stats APIs.

Also add an experiment to be used in a subsequent PR.

PiperOrigin-RevId: 647730284