[benchmark] Add a scalable_unrestricted benchmark category.
The following benchmark hangs on 8-core machines, but passes on 30-core
machines:

```
cpp-protobuf-async-client-unary-1channel-64wide-128breq-8mbresp-insecure
```

This change creates a `scalable_unrestricted` category in addition to
the `scalable` category typically run in CI, and applies it to all
`scalable` benchmarks except the one that fails on 8-core machines.

This change also changes the set of benchmarks run on 8-core machines to
use the `scalable_unrestricted` category. Benchmarks run on 30-core
machines remain unchanged.