[export] Allow http proxy on a unix domain socket.
Supports either using the socket address as the authority (percent escaped);
or using an empty authority, and the socket path as the path.

following the convention by https://stackoverflow.com/questions/14771172/http-over-af-unix-http-connection-to-unix-socket#comment107156213_14771218

The intended use case is also similar (avoiding an extra local port in an inverting proxy set up)

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/653702831](http://cl/653702831)

PiperOrigin-RevId: 653702831