[HTTP client] pass URI to override functions (#37540)
Prior to this PR, the host and path were passed to the override functions, but the query params were unavailable to them.  I have replaced the separate `host` and `path` parameters with a single parameter that passes the URI in as a const reference, which provides access to the query params.

While I was at it, I also changed the PUT and POST override methods to pass in the body as a string_view, which is more ergonomic for tests.

Closes #37540

COPYBARA_INTEGRATE_REVIEW=https://github.com/grpc/grpc/pull/37540 from markdroth:httpcli_override_uri 902b0d2097180f53cdc3142f6416184b0c6ebb2a
PiperOrigin-RevId: 665950188