[xDS] validate Audience cluster metadata (#37566)
Add validation of the `Audience` cluster metadata type, as per gRFC A83 (https://github.com/grpc/proposal/pull/438).

I had previously changed the metadata to be represented as JSON in #37468.  However, while working on the GCP Authentication filter implementation, I realized that that's not an ideal representation, because it would have required us to validate the JSON on a per-RPC basis, which would be bad for performance.  So I've changed the representation of metadata to be an abstract type, and we now store the `Audience` metadata as a simple string.  I've also moved metadata into its own type with its own validation code, so that in the future we can use it in places other than CDS (many xDS resource types have metadata fields).

While I was at it, I also add some helper functions for validating the `UInt32Value` and `UInt64Value` wrapper protos.

Closes #37566

PiperOrigin-RevId: 668281729