[Ruby] Reduce log when starting a server
Since v1.65.0, GRPC always output log when starting a server.

```
WARNING: All log messages before absl::InitializeLog() is called are written to STDERR
I0000 00:00:1724995922.684820  513990 server.cc:1797] Completion queue of type 1 is being registered as a server-completion-queue
```

According to the e7e38da237 and comment, it seems that this log
always output when using Ruby.
So I changed it to use `VLOG` as same as #37296.