[grpclb_e2e_test] fix a couple of flakes (#37677)
One flake is in the `UsePickFirstChildPolicy` test, where PF happens to choose the second backend instead of the first one.  This can happen when the connection attempts are run in parallel due to happy eyeballs and the second one happens to finish before the first one.  Example failure:

https://btx.cloud.google.com/invocations/a48ad3f4-c3fb-4cc7-9af4-dec027504f4a/targets/%2F%2Ftest%2Fcpp%2Fend2end:grpclb_end2end_test@poller%3Dpoll;config=e15de0119d30933c31c54fd53f0490b0130f57af5825b09748361d587df0a7d0/tests

I fixed this by changing the test to require that all traffic goes to one of the two backends instead of requiring that it is the first one.

The other flake is in the `SwapChildPolicy` test, where the traffic is not being evenly distributed to the backends under round robin.  Example failure:

https://btx.cloud.google.com/invocations/57ac3d4c-ce2a-45be-8147-0cbed8dd8786/targets/%2F%2Ftest%2Fcpp%2Fend2end:grpclb_end2end_test@experiment%3Dwork_serializer_dispatch;config=56f5b09615e325097b100b58c41171656571290519a83c5d89a6067ef0283d46/tests

I'm honestly not sure what's causing this, but I don't think the test actually needs to verify the behavior of round_robin in the first place; it's sufficient to just check that traffic is going to both backends to know that we've switched from PF to RR.  So I've simply removed the additional checks here.

Closes #37677

COPYBARA_INTEGRATE_REVIEW=https://github.com/grpc/grpc/pull/37677 from markdroth:grpclb_e2e_test_flake 48b78ca8adb109c0023c7ff0ddbc724320ebb5b9
PiperOrigin-RevId: 673525103