[testing] reduce number of sockets in windows socket-use-after-close detection loop (#37669)
This test has been flaking for a while with a WSAEACCESS error on the `bind` call.

Change the loop to only create on socket at a time (on Windows) to rule out something windows-specific is not liking the fact that we are opening multiple listen sockets on the same port.

Closes #37669

COPYBARA_INTEGRATE_REVIEW=https://github.com/grpc/grpc/pull/37669 from apolcyn:change_loop ffa105ba462af814782e352c49823bef89598741
PiperOrigin-RevId: 675172803