Enable vsock support for Android
The current code tries to enable vsock on Android when a certain

a) Linux version is detected
b) and a certain libc is present.

For a), it requires that <linux/version.h> is included.

For b), AF_VSOCK is included since NDK version 14.

This commit fixes both issues by including the needed Linux header and
detecting the NDK_MAJOR version.

For context, AOSP, has it always enabled:
https://android.googlesource.com/platform/external/grpc-grpc/+/06fb97455fd5336cebfa9a1eb44d4fac6d054608%5E%21/src/core/lib/iomgr/port.h

But this patch compiles with older NDK version, in case someone is still
using them.