Don't grab GIL from C-Core during shutdown (retry with fix)
This is a retry of #37762 with a fix. We no longer hold the C mutex and GIL at the same to avoid a deadlock.

MetadataPluginCallCredentials passes a function pointer to C-core to be called during destruction. This function grabs GIL which may race between GIL destruction during process shutdown. Since GIL destruction happens after Python's exit handlers, we cana mark that Python is shutting down from an exit handler and don't grab GIL in this function afterwards using a C mutex.