[xDS e2e tests] use real xDS protos
This removes all xDS protos except for 5 of them that have services.  We still have some limitations in our internal build system that make it hard to use the real xDS protos for those files, but we're now using the real xDS protos for the rest.

(Note: discovery.proto is actually a special case.  While it does have services, we don't actually use those services, so that's not the reason we need a copy of this file.  Unfortunately, the xDS BUILD files group discovery.proto into the same build target as ads.proto, which has services that we actually use, thus requiring us to have our own copy.  This means that depending on the real discovery.proto causes us to also depend on the real ads.proto, which causes a conflict in the protobuf registry by linking two copies of ads.proto.  However, we *are* using the real discovery.proto in unit tests, which do not depend on ads.proto.)

PiperOrigin-RevId: 693907782