[CI] Upgrade to Bazel 8 (#38254)
Bazel 8 was released in Dec 2024 and gRPC picks this up as gRPC is expected to support the latest Bazel version. This upgrade needs many necessary changes to deal with its breaking changes. Those changes were made in separate PRs so that this very change stays small and simple. In addition to obvious mechanical changes, a noteworthy change is made; adding `common --enable_workspace=true` to the `bazel.rc` file. This re-enables workspace functionality, which was disabled by default in Bazel 8. This is a temporary measure and will be removed once gRPC completes its migration to bzlmod later this year.

All comments below are working logs that I've made and don't need to be read for the code review.

Closes #38254

PiperOrigin-RevId: 718488476