[Build] Minimal bzlmod support (#38456)
This is the first try to enable us to upload gRPC bzlmod when publishing a new gRPC C++ release while keeping using WORKSPACE.

- Based on gRPC bzlmod 1.69 ([link](https://github.com/bazelbuild/bazel-central-registry/tree/main/modules/grpc/1.69.0))
  - Removed `grpc-java` dependency as it is not needed.
  - Removed `disable-layering-check.patch` which was added as a workaround for mac layering issue.
  - Removed the patch for `bazel/python_rules.bzl`. This will be handled by https://github.com/grpc/grpc/pull/36843 and https://github.com/grpc/grpc/pull/35666.
- Added a new test `bazel_build_with_bzlmod_linux` to make sure that gRPC can be built with bzlmod.

Currently, there's no mechanism to automatically enforce dependency consistency between the Bazel workspace and bzlmod. This remains an unresolved issue, though it may become less relevant as the workspace is phased out in favor of bzlmod.

Closes #38456

PiperOrigin-RevId: 719504680