[pick_first] fix bug that caused us to stop triggering connection attempts (#38714)
The bug was triggered in a fairly rare sequence of events: the subchannel needed to report TRANSIENT_FAILURE as its initial state but then be in IDLE state by the time that Happy Eyeballs got to it.  Because we had recorded that the subchannel already saw TRANSIENT_FAILURE state before Happy Eyeballs got to it, it was not considered a new failure, which broke the logic we use to trigger exiting from the initial Happy Eyeballs pass and continue connecting.  The fix is to not record that the subchannel saw TRANSIENT_FAILURE state until we get to it in the Happy Eyeballs pass.

b/394267714

Closes #38714

COPYBARA_INTEGRATE_REVIEW=https://github.com/grpc/grpc/pull/38714 from markdroth:pick_first_happy_eyeballs_bugfix 97eb11f9ef8389d95cc03540f44b5989a0071346
PiperOrigin-RevId: 726236599