Revert "[promises] Add an optimizer for `Seq`, `TrySeq` (#38445)" (#38748)
This reverts commit a7a650a74e6106ecbccd3f28829c542722f766d6.

Original PR breaks #38286 and I think the breakage is in the PR -- in a construct like:

```
TrySeq(
  /*a*/ []() -> Promise<absl::Status>,
  /*b*/ []() -> Promise<StatusFlag>,
  /*c*/ []() -> Promise<absl::Status>)
```

the original seq_state code would try to promote the intermediate `StatusFlag` to an `absl::Status` which is great - however with the optimization, if `b` is instantaneous we try to demote the result of `a` to `StatusFlag`, losing information.

I'll need to consider how to change this optimization to preserve the final output type before rolling forward.

Closes #38748

COPYBARA_INTEGRATE_REVIEW=https://github.com/grpc/grpc/pull/38748 from ctiller:rb1 743eb33bdf7af0dacd7bef50d12411c30f2198ea
PiperOrigin-RevId: 727109120