[promises] Handle immediate-value promise factories closer to as expected (#38756)
Builds on #38745

When we have a promise factory of the form:

`[]() { blah(); return 7; };`

We currently rewrite it to a promise factory of

```
[]() {
  return []() -> Poll<int> { blah(); return 7; }
}
```

which is correct, but unintuitive -- the delayed execution of blah() after the factory execution is surprising, because if we make the small change to rewrite the original to:

`[]() { blah(); return Immediate(7); }`

then the promise factory is instead:

```
[]() {
  blah();
  return []() { return 7; };
}
```

and at least my intuition says that this should be the form of the original rewrite too. This change makes that so.

Closes #38756

COPYBARA_INTEGRATE_REVIEW=https://github.com/grpc/grpc/pull/38756 from ctiller:noup2 751843ae627aeabd909ea3c545c762dc08c0b9ba
PiperOrigin-RevId: 728243208