Fix race condition during pre-fork hook leading to assertion error
This commit fixes a race condition inside the Executor class which can
lead to assertion errors (joining an uninitialized Thread) or deadlocks.
The race exists between `SetThreading(false)` made during a
`grpc_shutdown_internal_locked` call and the `Enqueue` method which may
be called whenever a callback is registered. More details can be found
under the following issue comment
https://github.com/grpc/grpc/issues/38251#issuecomment-2685537047. The
TL;DR is that the thread state lock is not held continuously in the
`Enqueue` method between the time to decide to spawn a new Thread and
the time to actually instantiate it. This means that it can decide to
spawn one (ie. `shutdown` is false), the `SetThreading(false)` then sets
the `shutdown` attribute to false but the `Enqueue` method still creates
the thread. This race can have two bad outcomes for the program: 1) the
`SetThreading(false)` method may call `Join` on a thread that hasn't yet
been initialized (`num_threads_` is not accessed with a lock held) or 2)
the newly created Thread is never cleanly closed leading to a deadlock
(setting `shutdown` to false is done prior to the race and is the only
way to close a thread). This commit addresses both outcomes by
preventing the `Enqueue` method from creating a thread when the
`shutdown` attribute is false. It is done by re-acquiring this lock
right before actually creating the thread.