Protect against multiple concurrent run of SetThreading(false)
The SetThreading method is not thread safe and although it seems
intended to be run only once during grpc_shutdown, it's possible to run
multiple during pre-fork. This commit addresses a situation where the
executor threads are joined more than once which is invalid as per the
implementation of Thread.