[grpc_error] add experiment to flatten errors (#38874)
This experiment introduces the following changes:
- There are no more hierarchical errors; when calling `StatusAddChild()`, we now instead merge the two errors into a single `absl::Status`.  The rules for this merging are:
  - Code: If the parent's code is UNKNOWN, we use the child's code; otherwise, we use the parent's code.
  - Message: We append the child's message to the parent's.
  - Payload: We copy all payload keys from the parent.  Then we copy all payload keys from the child that aren't already present in the parent.
- The `kRpcStatus` and `kGrpcMessage` properties are no longer encoded in the payload; instead, they are reflected in the top-level code and message of the `absl::Status`.
  - Note: Because `absl::Status` does not retain the message if the status is OK, this means that gRPC will no longer retain the message sent from the server if the status is OK.  This is a behavior change, but it is consistent with our long-term desire to replace `grpc::Status` with `absl::Status`.  It is also consistent with behavior of grpc-go.  (grpc-java does retain the message in this case but exposes it only to the interceptor API, not the application API.)
- This also simplifies the logic in `grpc_error_get_status()` for determining the status.  The only special case left there is the mechanism for computing the status code from the HTTP2 error and vice-versa: when returning the status code, we convert from the HTTP2 error only if the status code is UNKNOWN, and when returning the HTTP2 error, we convert from the status code only if it is something other than UNKNOWN.  We can work toward removing that last special case in a subsequent effort.
- The legacy `grpc_error` APIs in src/core/lib/iomgr/error.h are now simple wrappers around the newer APIs in src/core/util/status_helpers.h.  There are no more special behaviors in the wrappers, so the behavior is the same, no matter which API a caller uses.

Note that in general, we are assuming that status UNKNOWN is the least desirable status code to use: in almost all cases, if we have any other possible code to use, the other code is almost certainly a better choice.  The vast majority of cases where we use this code today are cases where the status is created via `GRPC_ERROR_CREATE()` without explicitly setting a status code, and all of those cases will eventually go away by changing callers to explicitly create an `absl::Status`.  We do have a number of cases (most notably in the EventEngine code, probably since it was copied from the old iomgr code) where we're explicitly returning UNKNOWN errors, but I would argue that all of those cases should be changed to return a more appropriate status code instead.  So if we do run into any problems where one of those statuses is getting changed to something else by the systematic assumption here, then (a) it would be a case where UNKNOWN was the wrong code anyway and (b) the easy fix will be to change the place where we generate the error to use a better status code, in which case it will not get overwritten.

There is a small risk that this change will result in changing the status reported for a given RPC, since the logic for aggregating children is not exactly the same as it was in the past.  In particular, the old logic does a depth-first search through the error tree to find one that has an explicit RPC status set, whereas this new logic aggregates the status when merging errors, which may not be done in a depth-first order, thus resulting in a different outcome.  However, there are only ~20 callers actually adding child statuses, so the risk of this doesn't seem huge.

The main goal of this change is to get us to a place where there are no more hierarchical errors and the logic for propagating errors up the stack is stable and consistent.  This means that we can now start to change individual callers to construct `absl::Status` directly rather than using any of the APIs in src/core/util/status_helpers.h or src/core/lib/iomgr/error.h -- and we can change the callers in any order without the risk of affecting the overall status computation.

Closes #38874

COPYBARA_INTEGRATE_REVIEW=https://github.com/grpc/grpc/pull/38874 from markdroth:error_flatten e05975386051061067b7d93edb40529587a42d3f
PiperOrigin-RevId: 740472859