[core-e2e] Fix no-logging logic (#39119)
This test is intended to verify that there's no logs on the ordinary RPC path - but not necessarily on the channel creation paths.

Since there's no way to know the origination of the log, we allow the first RPC to emit some spam, so long as subsequent ones do not.

This carve out was removed in #37177 and since then we've had a slow but continuous stream of additions to the exemption list.

Returning to the original logic in the hopes that we no longer need to maintain said list. In a follow-up PR we should trim the exemptions we've added once more.

Closes #39119

COPYBARA_INTEGRATE_REVIEW=https://github.com/grpc/grpc/pull/39119 from ctiller:no-logging e15b67a598972d369e2b0071b0e04e6b48cda1ba
PiperOrigin-RevId: 742943838