Return `absl::UnavailableError` when the DNS entry cannot be resolved
The reason for this change is that the `google::cloud::storage` Grpc client retries `UnavailableError` but not `UnknownError`. When a lot of processes are using Grpc on the same machine, we end up with DNS resolution issues, which is why it is important to ensure that the error is retried.