[ruby] Expose method to control the thread that initializes gRPC
If gRPC is lazily initialized on a different thread than the one that
calls fork it will raise.

If an app server is managing the fork calls
it also needs to be able to force the initialization of the threads
so that it can be sure both are done on the same thread.

This simply exposes a method to Ruby to force this initialization
so that apps don't have to resort to hacks like calling GRPC::Core::Channel.allocate.