[export] Split CoreConfiguration builders into two buckets: ephemeral and persistent.
We use builders for a variety of tasks - sometimes for testing, sometimes for adapting gRPC to novel environments, sometimes for once off special tasks early in a process lifetime.

Treating these as the same thing has started leading to some problems, so here I introduce a reset-proof registration for a subset of these tasks.

I'll be using this to register chaotic-good into the EndpointTransport registry so that it doesn't need to join the usual core configuration chain (and accidentally pull in protobuf to core).

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/748704078](http://cl/748704078)

PiperOrigin-RevId: 748704078