[Backport][v1.70.x][EventEngine] Fix busy loop in thread pool when shutting down (#39258) (#39317)
Backports https://github.com/grpc/grpc/pull/39258 to v1.70.x branch.

----

Potential alternative fix for https://github.com/grpc/grpc/pull/39240.

When the WorkStealingThreadPool is shutting down, the lifeguard thread
could end up in a busy loop while waiting for the pool to quiesce. In
some environments, this may prevent other threads from running in a
timely manner, which can cause the process to hang. This PR adds a short
delay on lifeguard thread's quiesce check, which grows exponentially to
a maximum of 1 second if the thread pool is idle.