Fix races with data source shutdown
When we start destroying a data source we'll first call the leaf destructor, then update the vtable, then remove ourselves from the BaseNode data source list.

When this occurs and we get a channelz query racing we will call into an invalid vtable.

Instead, in any derived type, forcibly remove ourselves at the start of the most derived destructor, and then proceed with destruction normally.

PiperOrigin-RevId: 754199982