[party] Add a JSON rendering
Use new promise JSON rendering in parties to allow querying the total party state on demand.

We use a notification and a background thread to ensure timely queries - note that this will not work from fuzzers.

The intended usage is to extricate data via channelz and the test postmortem analysis tooling.

PiperOrigin-RevId: 762286595