Fix reading of uninitialized bool (cancelled_) upon moving NextResult.
Not all constructors initialize it. By contract this should be fine, except NextResult is moveable.
The default move constructor will read the uninitialized `cancelled_`, triggering UB.

This fixes by initializing the value to be semantically correct. An alternative approach would be to implement move manually to skip that field when it is not relevant.