[export] [mpsc] Add a probe class
Allow external observers of the MPSC to observe the number of queued tokens.

Intended use case is for load balancing in CG where we can see the number of queued tokens outstanding and make decisions based upon that.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/767241410](http://cl/767241410)

PiperOrigin-RevId: 767241410