[xDS e2e test] change LongRunningRpc to use callback API (#39814)
This fixes a flake in the XdsRoutingClusterUpdateClustersWithPickingDelays test, which tests to see that RPCs are tied to the xDS cluster picked when the RPC is started.  The test was starting an RPC in a background thread, which it assumed was tied to the current config, and then changing the config in the main thread.  The flake occurred when the background thread did not actually manage to start the RPC before the main thread changed the config, so the RPC wound up using the new config instead of the original config.

To fix this, I have changed the test framework's `LongRunningRpc` implementation to use the callback API instead of starting a background thread.  This ensures that when a test starts the RPC, the RPC will actually be started immediately, which should both address the flake I describe above and avoid other such flakes that may occur in other tests.

While I was at it, I reordered the steps in the test to make it a little clearer and more debuggable, although those changes don't actually affect the test behavior.

Closes #39814

COPYBARA_INTEGRATE_REVIEW=https://github.com/grpc/grpc/pull/39814 from markdroth:xds_e2e_test_flake_fix e6e8fa9c9ac9e6ff9ea0afc5113d904845ccdcf0
PiperOrigin-RevId: 769762590