[export] [channelz-v2] Name pieces of data in an entity/trace
As I'm iterating it's pretty clear we want some kind of name against each category of data. I'm not sure we want to have unique names though, so I'm tending towards this idea of being able to attach named/typed things against the object of interest.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/773729563](http://cl/773729563)

PiperOrigin-RevId: 773729563